General: Added comments for things to come. Declarations moved outside of methods or loops.
SpaceColonyGame: I added a Boolean to mark weather or not the game is done.
Engine, List, and Factory: The compiler directed me to make a few modifications. First of all, the RandomEventList now imports arrayList, witch the compiler said it should have. It also, for some reason, said that an arrayList<randomEvent> could not be converted into an arrayList. So, since the two of them are supposed to be interchangeable, at least as far as the current program knows, i simply followed the chain of complaints, replacing the former with the latter, until the complaints stopped.

Engine now has a while loop of the actual process of turn-by-turn playing. Some of the entries are just comments to come back too later, but there is at least a blueprint.

The compiler compailned about the generateEvent() on RandomEventList not having a return, so I just had it return null. The nullPointerException you get is ENTIRELY ENTENTIONALL.